+++ lib/src/utils.dart
@@ -9,13 +9,15 @@ import 'package:flutter/services.dart';
 
 import 'mobile_ads.dart';
 
-/// Assert the running platform is supported.
-/// The supported platforms are: Android
+/// Make sure the running platform is supported.
+/// The currently supported platforms are:
+///   - Android
 void assertPlatformIsSupported() {
-  // Google Native ads are only supported in Android and iOS
+  // Google's AdMOB supports only Android and iOS
   assert(
     Platform.isAndroid || Platform.isIOS,
-    'The current platform does not support native ads. The platforms that support it are Android and iOS',
+    'The current platform does not support native ads. '
+    'The platforms that support it are Android and iOS',
   );
 
   // TODO: Support iOS
@@ -33,47 +35,41 @@ void assertMobileAdsIsInitialized() {
   );
 }
 
-/// Assert the Native or Banner Ad controller is attached and
-/// isn't disposed.
-void assertControllerIsAttached(bool attached) {
-  assert(attached, 'You can NOT use a disposed controller');
-}
-
-void assertControllerIsNotAttached(bool attached) {
-  assert(
-    !attached,
-    'This controller has already been attached to a native or banner ad. You need one controller for each',
-  );
-}
-
-void debugCheckIsTestId(String id, List<String> testIds) {
-  assert(id != null);
-  assert(testIds != null);
-  if (!testIds.contains(id) && kDebugMode)
-    print(
-      'It is highly recommended to use test ads in for testing instead of production ads'
-      'Failure to do so can lead in to the suspension of your account',
-    );
-}
-
 /// Assert the current version is supported.
 /// The min versions are:
 ///  - iOS: 9
 ///  - Android: 16 (19 for Native and Banner Ads)
-void assertVersionIsSupported() {
-  if (Platform.isAndroid)
+void assertVersionIsSupported([bool usePlatformView = true]) {
+  if (Platform.isAndroid) {
+    /// The min required version for Android is 16
     assert(
-      MobileAds.osVersion >= 19,
-      'Native and Banner Ads are not supported in versions before 19 because'
-      ' flutter only support platform views on Android 19 or greater.',
+      MobileAds.osVersion >= 16,
+      'The required version to use the AdMOB SDK is 16 or higher',
     );
-  else
+
+    /// The required version by flutter to use PlatformViews is
+    ///   - Hybrid composition: 19
+    ///   - Virtual display: 20
+    if (usePlatformView)
+      assert(
+        MobileAds.osVersion >= 19,
+        'Native and Banner Ads are not supported in versions before 19 because'
+        ' flutter only support platform views on Android 19 or greater.',
+      );
+  } else {
+    /// The min required version for iOS is 9
     assert(
       MobileAds.osVersion >= 9,
-      'The required version to use the AdMOB SDk is 9 or higher',
+      'The required version to use the AdMOB SDK is 9 or higher',
     );
+  }
 }
 
+/// The Ad Builder
+///
+/// Useful links:
+///   - https://github.com/bdlukaa/native_admob_flutter/wiki/Native-Ad-builder-and-placeholders#adbuilder
+///   - https://github.com/bdlukaa/native_admob_flutter/wiki/Banner-Ad-builder-and-placeholders#adbuilder
 typedef AdBuilder = Widget Function(BuildContext context, Widget child);
 
 /// Build the android platform view
@@ -140,6 +136,8 @@ class AdError {
   ///
   /// Global error codes:
   /// - Internal error (Something happened internally; for instance, an invalid response was received from the ad server): 0
+  ///
+  /// For more info, read the [documentation](https://github.com/bdlukaa/native_admob_flutter/wiki/Ad-error-codes)
   final int code;
 
   /// Gets an error message. For example "Account not approved yet".
@@ -161,7 +159,7 @@ class AdError {
     this.cause,
   });
 
-  /// Retrieve this from a json
+  /// Retrieve an [AdError] from a json
   factory AdError.fromJson(Map<String, dynamic> json) {
     return AdError(
       code: json['code'],
@@ -178,8 +176,36 @@ class AdError {
 mixin UniqueKeyMixin {
   final _key = UniqueKey();
 
-  /// The unique key of the class
+  /// The unique key of [this] class
   String get id => _key.toString();
+}
+
+mixin AttachableMixin {
+  bool _attached;
+
+  /// Check if the controller is attached to an Ad
+  bool get isAttached => _attached;
+
+  /// Attach the controller to an Ad
+  /// Throws an `AssertionException` if the controller is already attached.
+  ///
+  /// You should not call this function
+  @mustCallSuper
+  void attach([bool attach = true]) {
+    _assertControllerIsNotAttached(isAttached);
+    assert(attach != null);
+    _attached = attach;
+  }
+
+  /// Ensure the controller is not attached
+  void _assertControllerIsNotAttached(bool attached) {
+    if (attached == null) return;
+    assert(
+      !attached,
+      'This controller has already been attached to a native or banner ad.'
+      'You need one controller for each',
+    );
+  }
 }
 
 abstract class LoadShowAd<T> with UniqueKeyMixin {
@@ -192,7 +218,9 @@ abstract class LoadShowAd<T> with UniqueKeyMixin {
   MethodChannel channel;
 
   bool _disposed = false;
-  /// Check if the ad is loaded
+
+  /// Check if the ad is disposed. You can dispose the ad by calling
+  /// `ad.dispose()`
   bool get isDisposed => _disposed;
 
   @mustCallSuper